+++ src/main/java/com/sharpandrew/learnjava/graph/storage/DefaultEdgeTable.java
@@ -8,36 +8,31 @@ import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.ComparisonOperator;
 import com.amazonaws.services.dynamodbv2.model.Condition;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.Iterables;
-import com.sharpandrew.learnjava.graph.model.GraphPath;
-import com.sharpandrew.learnjava.graph.model.ImmutableGraphPath;
+import com.google.common.collect.ImmutableSet;
 import com.sharpandrew.learnjava.serverless.Environment;
 import java.util.List;
 import java.util.Set;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-public class DynamoDbPathDao implements PathDao {
-  private static DynamoDbPathDao instance = null;
+public final class DefaultEdgeTable implements EdgeTable {
+  private static DefaultEdgeTable instance = null;
 
   private final DynamoDBMapper mapper;
 
   @VisibleForTesting
-  DynamoDbPathDao(DynamoDBMapper mapper) {
+  DefaultEdgeTable(DynamoDBMapper mapper) {
     this.mapper = mapper;
   }
 
-  public static DynamoDbPathDao getInstance() {
+  public static DefaultEdgeTable getInstance() {
     if (instance == null) {
-      String tableName = Environment.getPathTableName();
+      String tableName = Environment.getEdgeTableName();
       DynamoDBMapperConfig config = DynamoDBMapperConfig.builder()
           .withTableNameOverride(
               DynamoDBMapperConfig.TableNameOverride.withTableNameReplacement(tableName))
           .build();
       DynamoDBMapper mapper = new DynamoDBMapper(
           AmazonDynamoDBClientBuilder.defaultClient(), config);
-      instance = new DynamoDbPathDao(mapper);
+      instance = new DefaultEdgeTable(mapper);
       return instance;
     } else {
       return instance;
@@ -45,39 +40,27 @@ public class DynamoDbPathDao implements PathDao {
   }
 
   @Override
-  public GraphPath get(String pathId) {
+  public Set<StorageEdge> get(String graphId) {
     DynamoDBScanExpression scanExpression = new DynamoDBScanExpression();
     scanExpression.addFilterCondition(
-        "pathId",
+        "graphId",
         new Condition()
             .withComparisonOperator(ComparisonOperator.EQ)
-            .withAttributeValueList(new AttributeValue(pathId)));
-    List<StoragePathEdge> matchingEdges = mapper.scan(StoragePathEdge.class, scanExpression);
-    List<Integer> vertices = matchingEdges.stream()
-        .map(StoragePathEdge::getStartVertex)
-        .collect(Collectors.toList());
-    vertices.add(Iterables.getLast(matchingEdges).getEndVertex());
-    return ImmutableGraphPath.builder()
-        .verticesStartToFinish(vertices)
-        .build();
+            .withAttributeValueList(new AttributeValue(graphId)));
+    List<StorageEdge> matchingEdges = mapper.scan(StorageEdge.class, scanExpression);
+    return ImmutableSet.<StorageEdge>builder().addAll(matchingEdges).build();
   }
 
   @Override
-  public String post(GraphPath graphPath) {
-    String pathId = UUID.randomUUID().toString();
-    Set<StoragePathEdge> storagePathEdges = IntStream
-        .range(0, graphPath.verticesStartToFinish().size() - 1)
-        .boxed()
-        .map(place -> {
-          StoragePathEdge storagePathEdge = new StoragePathEdge();
-          storagePathEdge.setPathId(pathId);
-          storagePathEdge.setPlace(place);
-          storagePathEdge.setStartVertex(graphPath.verticesStartToFinish().get(place));
-          storagePathEdge.setEndVertex(graphPath.verticesStartToFinish().get(place + 1));
-          return storagePathEdge;
-        })
-        .collect(Collectors.toSet());
-    List<DynamoDBMapper.FailedBatch> failedBatches = mapper.batchSave(storagePathEdges);
+  public Set<StorageEdge> getAll() {
+    List<StorageEdge> allEdges = mapper.scan(
+        StorageEdge.class, new DynamoDBScanExpression());
+    return ImmutableSet.<StorageEdge>builder().addAll(allEdges).build();
+  }
+
+  @Override
+  public void post(Set<StorageEdge> edges) {
+    List<DynamoDBMapper.FailedBatch> failedBatches = mapper.batchSave(edges);
     if (!failedBatches.isEmpty()) {
       StringBuilder stringBuilder = new StringBuilder();
       stringBuilder.append("Failed to save some batches:\n");
@@ -88,8 +71,12 @@ public class DynamoDbPathDao implements PathDao {
         stringBuilder.append(batch.getException());
       });
       throw new RuntimeException(stringBuilder.toString());
-    } else {
-      return pathId;
     }
+  }
+
+  @Override
+  public void delete(String graphId) {
+    Set<StorageEdge> edgesToDelete = get(graphId);
+    mapper.batchDelete(edgesToDelete);
   }
 }
