+++ src/main/java/com/sharpandrew/learnjava/graph/storage/PathTable.java
@@ -2,7 +2,7 @@ package com.sharpandrew.learnjava.graph.storage;
 
 import com.sharpandrew.learnjava.graph.model.GraphPath;
 
-public interface PathDao {
+public interface PathTable {
   GraphPath get(String pathId);
 
   /** Returns the path's ID. */
